refactor for scheduler, also pdf (not workig)
