+++ src/components/schedule/utils/constants.js
@@ -1,0 +1,7 @@
+
+export const DAYS = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri'];
+
+export const TIMES = [
+  '', '7 AM', '8 AM', '9 AM', '10 AM', '11 AM', '12 PM',
+  '1 PM', '2 PM', '3 PM', '4 PM', '5 PM', '6 PM', '7 PM', '8 PM'
+];
